rpsls_game_api/requirements.txt
black==22.8.0
fastapi==0.115.4
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.32.0
